Working EdDSA with tests.

+++ jostle/src/main/java/org/openssl/jostle/jcajce/provider/ed/EDServiceJNI.java
@@ -28,7 +28,7 @@ public class EDServiceJNI implements EDServiceNI
     public native int ni_initSign(long reference, long keyRef, String name, byte[] context, int contextLen, RandSource randSource);
 
     @Override
-    public native long ni_sign(long reference, byte[] sig, int i, RandSource randSource);
+    public native int ni_sign(long reference, byte[] sig, int i, RandSource randSource);
 
     @Override
     public native int ni_initVerify(long reference, long keyRef, String name, byte[] context, int contextLen);
